Added a registry to record run dates for harvester
This is an optional utility that can be used directly by harvesters to
log a run date and retrieve the most recent run date.  It shares the
database used by the HarvestRegistry.  To avoid confusion in the library
names, the original HarvestRegistry has been renamed to
HarvestRecordRegistry, and the new library is HarvestRunRegistry.