Added a registry to record run dates for harvester (#4)
This change is to add an optional utility that can be used directly by harvester apps to log a new run date and retrieve the most recent run date. It shares the database connection with the HarvestRegistry. To avoid confusion in the library names, the original HarvestRegistry has been renamed to HarvestRecordRegistry, and the new class is HarvestRunRegistry.