Minor fixes to DiSCO update logic
The way isLatest() was being used caused DiSCOs never to update, partly
because of the ambiguity of what HarvestRecordStatus.isLatest means -
whether it means the record in the registry isLatest, or the new record
isLatest.  Previously isLatest() was set to true when the harvested
record info passed in was newer than the the DiSCO version in the
registry.  Now, isLatest has been changed to isUpToDate(), and it is
true when the record in the registry is up to date.  If the registry is
up to date, the newly harvested record info does not need to be
processed. New logic reflects this. An integration tests demonstrates
the behavior.