Optional basic duplicate checker that compares 2 DiSCOs
Provides optional validation to check whether two DiSCOs are the same.
This can be used to prevent DiSCO updates that repeatedly POST the same
unchanged DiSCO to RMap as a new version.  The DiSCO URI is ignored in
this comparison. The comparison is fairly rudimentary - it has been
manually tested, but needs unit tests. It will not work if there are any
BNodes in the graph apart from the DiSCO URI since RMap removes these
and replaces them with persistent identiers.