Fixes to DiSCO update logic (#6)
The way isLatest() was being used caused DiSCOs never to update, partly
because of the ambiguity of what HarvestRecordStatus.isLatest means -
whether it means the record in the registry isLatest, or the new record
isLatest.  Previously isLatest() was set to true when the harvested
record info passed in was newer than the the DiSCO version in the
registry.  Now, isLatest has been changed to isUpToDate(), and it is
true when the record in the registry is up to date.  If the registry is
up to date, the newly harvested record info does not need to be
processed. New logic reflects this. An integration tests demonstrates
the behavior.

* Fixed tests to reflect recent changes

Specifically, the DiSCO URI for the registry is now taken from the body
of a 201 response, rather than the Location header, which includes the
API path. Also fixed a nullpointer for the skipUpToDateTest

* Corrected a comment in the test

* Fixed early countdown latch trigger in successfulSendTest

 The latch was being triggered before the record value is being set in successfulSendTest().